Switch to the DiscoverRunner test runner
