+++ opentreemap/treemap/tests/test_images.py
