--- opentreemap/treemap/tests/ui/plot_detail/__init__.py
@@ -1,3 +1,0 @@
-from delete import *  # NOQA
-from edit import *  # NOQA
-from add import *  # NOQA
