+++ opentreemap/treemap/tests/ui/plot_detail/uitest_delete.py
