+++ opentreemap/treemap/tests/ui/ui_test_urls.py
@@ -1,5 +1,9 @@
-from django.conf.urls import patterns, include, url
+# -*- coding: utf-8 -*-
+from __future__ import print_function
+from __future__ import unicode_literals
+from __future__ import division
 
+from django.conf.urls import patterns, include, url
 from django.http import HttpResponse
 
 from opentreemap import urls
